first try to support campuscard
